hal/dx12: move device and encoder into modules
